Added Line Renderering, including a new shader and new renderer. Modified a bit the way shaders are organised as well as renderers.

+++ src/com/thibsworkshop/voxand/debugging/Timing.java
@@ -12,6 +12,7 @@ public class Timing {
 
     //IMPORTANT: One record at a time for each method of each class, otherwise timing wont be correct
 
+    //TODO: Maybe change the List of data to an array of fixed size, like 100, to avoid deleting the first element of the list
     private static Map<String, Map<String, List<Long>>> data = new HashMap<>(); //The data
     private static Map<String, Map<String, Boolean>> enabled = new HashMap<>(); //Is the method enabled for recording?
 
@@ -73,7 +74,7 @@ public class Timing {
         NumberFormat nf = NumberFormat.getInstance();
         nf.setMaximumFractionDigits(precision);
         for(float val : get(name,method)){
-            System.out.println(nf.format((double)val/1000f));
+            System.out.println(nf.format(val/1000f));
         }
     }
 }
